Break out sending into separate threads; more socket wip

+++ hybrasyl/Objects/User.cs
@@ -531,7 +531,7 @@ namespace Hybrasyl.Objects
             
         }
 
-        public bool AssociateConnection(World world, IntPtr connectionId)
+        public bool AssociateConnection(World world, long connectionId)
         {
             World = world;
             Client client;
@@ -540,7 +540,7 @@ namespace Hybrasyl.Objects
             return true;
         }
 
-        public User(World world, IntPtr connectionId, string playername = "")
+        public User(World world, long connectionId, string playername = "")
         {
             World = world;
             Client client;
